feat: add executor custom  logic

+++ src/solana/trigger.rs
@@ -9,7 +9,9 @@ impl Trigger<TaskAccount> for TimestampTrigger {
             .duration_since(std::time::SystemTime::UNIX_EPOCH)
             .unwrap_or_default()
             .as_secs();
-        if current_time > task.end_at {
+        
+        // if current_time > task.end_at {
+        if current_time as i64 > task.end_at {
             return true;
         }
         false
